langchain-apps/pyproject.toml
[project]
name = "langchain-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "isort>=8.0.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.22",
]
